requirements.txt
docling
docling-ibm-models
mlx-vlm
PyYAML
beautifulsoup4

